Move "Fetch code data" to Map VC
