+++ WeatherApp/ViewControllers/Favorites/FavoritesViewController.swift
@@ -18,7 +18,7 @@ class FavoritesViewController: UIViewController {
         tableView.delegate = self
         tableView.dataSource = self
         presenter.view = self
-        presenter.fetchCoreData()
+       // presenter.fetchCoreData()
         setUI()
     }
